separate google map class on parking map activity
